refactor feature

+++ ewm-service/src/main/java/ru/practicum/ewm/user/dto/UserDto.java
@@ -17,6 +17,6 @@ public class UserDto {
     private String name;
     @NotNull
     private String email;
-    private Double rating;
+    private Long rating;
 
 }
